+++ src/index.js
@@ -2,19 +2,16 @@ const fs = require('fs');
 const jwt = require('jsonwebtoken');
 const joi = require('joi');
 const { validateCredentials } = require('./validate-credentials');
-const errorPrefix = 'Auth plugin: ';
 const TOKEN_EXPIRATION_MINUTES = 60;
 
 const optionsSchema = joi.object({
-  useHttp: joi.boolean().default(false),
   tokenExpirationMinutes: joi
     .number()
     .integer()
     .greater(4)
     .default(TOKEN_EXPIRATION_MINUTES),
-});
+}).unknown();
 
-let _useHttp;
 let _tokenExpirationMinutes;
 let _secret;
 let _userStoreFilePath;
@@ -25,46 +22,36 @@ let _userStoreFilePath;
  * @param {string}   userStoreFilePath - file path of user store JSON file
  * @param {object}   options
  * @param {integer}  options.tokenExpirationMinutes - number of minutes until token expires
- * @param {boolean}  options.useHttp - direct consumers of authenticationSpecifcation to use HTTP instead of HTTPS
  */
 function initAuthPlugin(secret, userStoreFilePath, options = {}) {
   // Throw error if user-store file does not exist
   const result = fs.existsSync(userStoreFilePath);
   if (!result) {
-    throw new Error(`${errorPrefix}${userStoreFilePath} not found`);
+    throw new Error(`${userStoreFilePath} not found`);
   }
   _secret = secret;
   _userStoreFilePath = userStoreFilePath;
 
   const {
     error,
-    value: { tokenExpirationMinutes, useHttp },
+    value: { tokenExpirationMinutes },
   } = optionsSchema.validate(options);
 
   if (error) {
-    throw new Error(`${errorPrefix}${error.details[0].message}`);
+    throw new Error(error.details[0].message);
   }
 
-  _useHttp = useHttp;
   _tokenExpirationMinutes = tokenExpirationMinutes;
 
   return {
     type: 'auth',
-    authenticationSpecification,
     authenticate,
     authorize,
+    name: 'file-based-auth-store',
+    version: require('../package.json').version
   };
 }
 
-/**
- * Return "authenticationSpecification" object for use in output-services
- * @returns {object}
- */
-function authenticationSpecification() {
-  return {
-    useHttp: _useHttp,
-  };
-}
 
 /**
  * Authenticate a user's submitted credentials
@@ -72,9 +59,17 @@ function authenticationSpecification() {
  * @returns {Promise}
  */
 async function authenticate(req) {
+  const expires = Date.now() + _tokenExpirationMinutes * 60 * 1000;
   const { query = {}, body = {} } = req;
-  const { username, password } = {...query, ...body};
+  const { username, password, token } = { ...query, ...body };
 
+  if (token) {
+    const { sub } = decodeToken(token);
+    return {
+      token: createToken(sub, expires),
+      expires,
+    };
+  }
 
   // Validate user's credentials
   const valid = await validateCredentials(
@@ -90,12 +85,9 @@ async function authenticate(req) {
   }
 
   // Create access token and wrap in response object
-  const expires = Date.now() + _tokenExpirationMinutes * 60 * 1000;
+  
   return {
-    token: jwt.sign(
-      { exp: Math.floor(expires / 1000), sub: username },
-      _secret,
-    ),
+    token: createToken(username, expires),
     expires,
   };
 }
@@ -106,7 +98,10 @@ async function authenticate(req) {
  * @returns {Promise}
  */
 async function authorize(req) {
-  let token = req.query?.token || req.headers['authorization'];
+  const { query = {}, body = {} } = req;
+  const headerToken = req.headers['authorization'] ?? req.headers['authorization'].replace(/^Bearer /, '');
+  const params = { ...query, ...body };
+  const token = headerToken || params.token;
 
   if (!token) {
     let err = new Error('No authorization token.');
@@ -114,6 +109,10 @@ async function authorize(req) {
     throw err;
   }
   // Verify token with async decoded function
+  return decodeToken(token);
+}
+
+async function decodeToken(token) {
   try {
     const decoded = await jwt.verify(token, _secret);
     return decoded;
@@ -121,6 +120,10 @@ async function authorize(req) {
     err.code = 401;
     throw err;
   }
+}
+
+function createToken(sub, expires) {
+  return jwt.sign({ exp: Math.floor(expires / 1000), sub }, _secret);
 }
 
 module.exports = initAuthPlugin;